+++ kdv_equation.h
@@ -132,7 +132,7 @@ double du_xxx( Vector *u, int i, double dx_3 );
 /***************************************************************/
 void simple_sec( Vector *u0, double dx, double M, double loc);
 
-void print_timing( double numSteps, double dt, double writeEvery,  
+void print_timing( double numSteps, double dt, int writeEvery,  
     double left, double right, double dx, char* file_name );
